Source Code for Scraper Project (beta)
this project includes complete set of code. it finds all links and
creates the xml for each link. Any runtime errors are log into a file
named "ErrorLog.txt". there are no current issues with this code. it
still needs to spot checked for any issues in the xml format.